clean up handler, move the multi goals to its own node. name planner and controller options by their class
