+++ cone_association/cone_association/node_trackdrive_handler_multiple.py
@@ -6,16 +6,19 @@ from tf2_ros.buffer import Buffer
 from tf2_ros.transform_listener import TransformListener
 
 import rclpy
+from rclpy.action import ActionClient
 
 from driverless_msgs.msg import Shutdown, State
 from geometry_msgs.msg import PoseWithCovarianceStamped, PoseStamped
 from std_msgs.msg import UInt8, Bool
+from nav2_msgs.action import NavigateThroughPoses
 
 from driverless_common.shutdown_node import ShutdownNode
 
 class TrackdriveHandler(ShutdownNode):
     mission_started = False
     crossed_start = False
+    sent_goal = False
     laps = 0
     last_lap_time = 0.0
     last_x = 0.0
@@ -26,7 +29,7 @@ class TrackdriveHandler(ShutdownNode):
     def __init__(self):
         super().__init__("trackdrive_logic_node")
 
-        self.declare_parameter("start_following", False)
+        self.declare_parameter("start_following", True)
 
         self.create_subscription(State, "system/as_status", self.state_callback, 1)
         self.create_subscription(Bool, "system/map_status", self.map_callback, 1)
@@ -39,7 +42,11 @@ class TrackdriveHandler(ShutdownNode):
         self.shutdown_pub = self.create_publisher(Shutdown, "system/shutdown", 1)
         self.lap_trig_pub = self.create_publisher(UInt8, "system/laps_completed", 1)
         self.init_pose_pub = self.create_publisher(PoseWithCovarianceStamped, "/initialpose", 1)
-        self.goal_pose_pub = self.create_publisher(PoseStamped, "/goal_pose", 1)
+
+        # actions
+        self.nav_through_poses_client = ActionClient(self,
+                                                     NavigateThroughPoses,
+                                                     'navigate_through_poses')
 
         if self.get_parameter("start_following").value:
             # start at lap 1
@@ -104,7 +111,7 @@ class TrackdriveHandler(ShutdownNode):
             self.get_logger().info(f"Lap {self.laps} completed")
 
             # publish initial pose on first lap
-            if self.laps == 1:
+            if self.laps == 1 and not self.sent_goal:
                 init_pose_msg = PoseWithCovarianceStamped()
                 init_pose_msg.header.stamp = track_to_base.header.stamp
                 init_pose_msg.header.frame_id = "track"
@@ -115,11 +122,17 @@ class TrackdriveHandler(ShutdownNode):
                 # cov diag to square
                 self.init_pose_pub.publish(init_pose_msg)
 
-            goal_pose_msg = PoseStamped()
-            goal_pose_msg.header.stamp = track_to_base.header.stamp
-            goal_pose_msg.header.frame_id = "track"
-            goal_pose_msg.pose.position.x = self.goal_offet - (self.laps * 0.1)
-            self.goal_pose_pub.publish(goal_pose_msg)               
+                goal_poses = []
+                for i in range(10):
+                    goal_pose_msg = PoseStamped()
+                    goal_pose_msg.header.frame_id = "track"
+                    goal_pose_msg.pose.position.x = self.goal_offet - (i * 0.1)
+                    goal_poses.append(goal_pose_msg)
+            
+                self.go_through_poses(goal_poses)
+                # self.goal_pose_pub.publish(goal_pose_msg)
+                self.sent_goal = True
+                self.get_logger().info("Trackdrive mission goal sent")
             
             self.last_x = track_to_base.transform.translation.x
             self.last_lap_time = time.time()
@@ -131,6 +144,29 @@ class TrackdriveHandler(ShutdownNode):
             shutdown_msg = Shutdown(finished_engage_ebs=True)
             self.shutdown_pub.publish(shutdown_msg)
 
+    def go_through_poses(self, poses: PoseStamped):
+        # Sends a `NavThroughPoses` action request
+        self.get_logger().info("Waiting for 'NavigateThroughPoses' action server")
+        while not self.nav_through_poses_client.wait_for_server(timeout_sec=1.0):
+            self.get_logger().info("'NavigateThroughPoses' action server not available, waiting...")
+
+        goal_msg = NavigateThroughPoses.Goal()
+        goal_msg.poses = poses
+
+        self.get_logger().info('Navigating with ' + str(len(goal_msg.poses)) + ' goals.' + '...')
+        send_goal_future = self.nav_through_poses_client.send_goal_async(goal_msg)
+        # rclpy.spin_until_future_complete(self, send_goal_future)
+        # while not send_goal_future.done():
+        #     self.get_logger().info('Waiting for NavigateThroughPoses action to complete...')
+        #     time.sleep(1)
+        # self.goal_handle = send_goal_future.result()
+
+        # if not self.goal_handle.accepted:
+        #     self.error('Goal with ' + str(len(poses)) + ' poses was rejected!')
+        #     return False
+
+        # self.result_future = self.goal_handle.get_result_async()
+        # return True
 
 def main(args=None):
     rclpy.init(args=args)
