+++ qutms_nav2/config/nav2_params.yaml
@@ -146,8 +146,8 @@ local_costmap:
 planner_server:
   ros__parameters:
     expected_planner_frequency: 20.0
-    planner_plugins: ["GridBased"]
-    GridBased:
+    planner_plugins: ["SmacPlannerHybrid"]
+    SmacPlannerHybrid:
       plugin: "nav2_smac_planner/SmacPlannerHybrid"
       tolerance: 0.5                      # tolerance for planning if unable to reach exact pose, in meters
       downsample_costmap: false           # whether or not to downsample the map
@@ -183,7 +183,7 @@ controller_server:
     failure_tolerance: 0.3
     progress_checker_plugin: "progress_checker"
     goal_checker_plugins: ["general_goal_checker"] # "precise_goal_checker"
-    controller_plugins: ["FollowPath"]
+    controller_plugins: ["RegulatedPurePursuit"]
 
     # Progress checker parameters
     progress_checker:
@@ -197,7 +197,7 @@ controller_server:
       xy_goal_tolerance: 0.01
       yaw_goal_tolerance: 0.01
     # Controller parameters
-    FollowPath:
+    RegulatedPurePursuit:
       plugin: "nav2_regulated_pure_pursuit_controller::RegulatedPurePursuitController"
       desired_linear_vel: 8.0  # The desired maximum linear velocity to use.
       lookahead_dist: 4.0  # The lookahead distance to use to find the lookahead point.
